requirements.txt
python-telegram-bot==21.6
psutil==5.9.8
PyYAML==6.0.2
python-dotenv==1.0.1
requests==2.32.3
aiohttp==3.9.5
